requirements.txt
uvicorn[standard]
psycopg2-binary
fastapi